+++ TestGame/DemoAnimatorComponent.h
@@ -48,7 +48,7 @@ public:
 
 	TESTGAME_API virtual int Load() override;
 
-	TESTGAME_API virtual void Update(float deltaTime) noexcept override;
+	TESTGAME_API virtual void Update(double deltaTime) noexcept override;
 
 	TESTGAME_API virtual void Unload() override;
